Reorganize things to make it easier to test code fixes
