+++ MonoDevelop.MSBuild.Tests/Helpers/MSBuildDocumentTest.Diagnostics.cs
@@ -1,0 +1,108 @@
+// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+#nullable enable
+
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading;
+
+using Microsoft.Extensions.Logging;
+
+using MonoDevelop.MSBuild.Analysis;
+using MonoDevelop.MSBuild.Language;
+using MonoDevelop.MSBuild.Schema;
+
+using MonoDevelop.Xml.Tests;
+
+using NUnit.Framework;
+
+namespace MonoDevelop.MSBuild.Tests;
+
+partial class MSBuildDocumentTest
+{
+	public static IList<MSBuildDiagnostic> GetDiagnostics (
+		string source,
+		out MSBuildRootDocument parsedDocument,
+		ICollection<MSBuildAnalyzer>? analyzers = null,
+		bool includeCoreDiagnostics = false,
+		ILogger? logger = null,
+		MSBuildSchema? schema = null,
+		MSBuildRootDocument? previousDocument = null,
+		CancellationToken cancellationToken = default
+		)
+	{
+		// internal errors should cause test failure
+		logger ??= TestLoggerFactory.CreateTestMethodLogger ().RethrowExceptions ();
+
+		parsedDocument = GetParsedDocument (source, logger, schema, previousDocument, cancellationToken);
+
+		var analyzerDriver = new MSBuildAnalyzerDriver (logger);
+
+		if (analyzers != null && analyzers.Count > 0) {
+			analyzerDriver.AddAnalyzers (analyzers);
+		} else if (!includeCoreDiagnostics) {
+			throw new ArgumentException ("Analyzers can only be null or empty if core diagnostics are included", nameof (analyzers));
+		}
+
+		var actualDiagnostics = analyzerDriver.Analyze (parsedDocument, includeCoreDiagnostics, cancellationToken);
+
+		return actualDiagnostics ?? [];
+	}
+
+	public static void VerifyDiagnostics (string source, MSBuildAnalyzer analyzer, params MSBuildDiagnostic[] expectedDiagnostics)
+		=> VerifyDiagnostics (source, out _, new[] { analyzer }, false, false, null, expectedDiagnostics);
+
+	public static void VerifyDiagnostics (string source, ICollection<MSBuildAnalyzer> analyzers, bool includeCoreDiagnostics, params MSBuildDiagnostic[] expectedDiagnostics)
+		=> VerifyDiagnostics (source, out _, analyzers, includeCoreDiagnostics, false, null, expectedDiagnostics);
+
+	public static void VerifyDiagnostics (
+		string source,
+		ICollection<MSBuildAnalyzer>? analyzers = null,
+		bool includeCoreDiagnostics = false,
+		bool checkUnexpectedDiagnostics = false,
+		MSBuildSchema? schema = null,
+		MSBuildDiagnostic[]? expectedDiagnostics = null,
+		ILogger? logger = null,
+		MSBuildRootDocument? previousDocument = null
+		)
+		=> VerifyDiagnostics (source, out _, analyzers, includeCoreDiagnostics, checkUnexpectedDiagnostics, schema, expectedDiagnostics, logger, previousDocument);
+
+	public static void VerifyDiagnostics (
+		string source,
+		out MSBuildRootDocument parsedDocument,
+		ICollection<MSBuildAnalyzer>? analyzers = null,
+		bool includeCoreDiagnostics = false,
+		bool checkUnexpectedDiagnostics = false,
+		MSBuildSchema? schema = null,
+		MSBuildDiagnostic[]? expectedDiagnostics = null,
+		ILogger? logger = null,
+		MSBuildRootDocument? previousDocument = null
+		)
+	{
+		var actualDiagnostics = GetDiagnostics (source, out parsedDocument, analyzers, includeCoreDiagnostics, logger, schema, previousDocument);
+
+		foreach (var expectedDiag in expectedDiagnostics ?? []) {
+			bool found = false;
+			for (int i = 0; i < actualDiagnostics.Count; i++) {
+				var actualDiag = actualDiagnostics[i];
+				if (actualDiag.Descriptor == expectedDiag.Descriptor && actualDiag.Span.Equals (expectedDiag.Span)) {
+					Assert.That (actualDiag.Properties ?? Enumerable.Empty<KeyValuePair<string, object>> (),
+						Is.EquivalentTo (expectedDiag.Properties ?? Enumerable.Empty<KeyValuePair<string, object>> ())
+						.UsingDictionaryComparer<string, object> ());
+					found = true;
+					actualDiagnostics.RemoveAt (i);
+					break;
+				}
+			}
+			if (!found) {
+				Assert.Fail ($"Did not find expected diagnostic {expectedDiag.Descriptor.Id}@{expectedDiag.Span.Start}-{expectedDiag.Span.End}");
+			}
+		}
+
+		if (checkUnexpectedDiagnostics && actualDiagnostics.Count > 0) {
+			Assert.Fail ($"Found unexpected diagnostics: {string.Join ("", actualDiagnostics.Select (diag => $"\n\t{diag.Descriptor.Id}@{diag.Span.Start}-{diag.Span.End}"))}");
+		}
+	}
+}
